requirements.txt
Pillow
python-multipart
ffmpeg
insightface
onnxruntime-gpu
numpy==1.23.5
opencv-python
opencv-contrib-python

fastapi
uvicorn[standard]
